==> 3. Tagging dataset.ipynb <==
import pandas as pd
import numpy as np
from scipy import spatial
import matplotlib.pyplot as plt
from sklearn.manifold import TSNE
import smote_variants as sv
from tqdm import tqdm
from urllib.parse import urlparse
import requests
from collections import Counter
import pandas as pd
import os
from flashtext.keyword import KeywordProcessor

df1=pd.read_csv("dataset/street list.csv")
df2=pd.read_csv("dataset/1smoteog.csv")
# =pd.read_csv("dataset/syncosinestreet.csv")
df3=pd.read_csv("dataset/3smotecos.csv")
df4=pd.read_csv("dataset/StreetListCleaned.csv")
oglist=list(df1['street'])
smote=list(df2['0'])
cossmote=list(df3['0'])
cleanog=list(df4['Words'])
print("Number of intial words: ",len(oglist))
print("Number of intial words cleaned: ",len(cleanog))
print("Number of synthetic words generated using smote: ",len(smote))
print("Number of synthetic words generated using improvised smote ",len(cossmote))

data=pd.read_csv('dataset/crawledcleaned.csv')
data

data.isnull().sum()

data.dropna(subset=['cleaned_website_text'],inplace=True)
data.isnull().sum()

from flashtext.keyword import KeywordProcessor
kp0=KeywordProcessor()
c=0
for word in oglist:
    kp0.add_keyword(word)
    c=c+1
print(c)
def percentage1(dum0,dumx):
    try:
        ans=float(dumx)/float(dum0)
        ans=ans*100
    except:
        return 0
    else:
        return ans
def find_category(x):
    try:
        Total_matches=kp0.extract_keywords(x)
        y0=len(Total_matches)
        n=len(x)
        per=percentage1(n,y0)
        if y0==0:
            matched_catkeywords='None'
        matched_catkeywords=Total_matches
        return y0,matched_catkeywords,per
    except:
        print(x)
Total=[]
words=[]
per=[]
df=data
for i in df['cleaned_website_text']:
    Total.append(find_category(i)[0])
    words.append(find_category(i)[1])
    per.append(find_category(i)[2])
df['no of matched keywords']=Total
df['words matched']=words
df['percent']=per
cat=[]
for i in per:
    if i>0:
        cat.append("drug")
    else:
        cat.append("non drug")
    
df['category']=cat
df['category'].value_counts()

df.to_csv("datawithoglist.csv")

kp0=KeywordProcessor()
c=0
for word in smote:
    kp0.add_keyword(word)
    c=c+1
print(c)
def percentage1(dum0,dumx):
    try:
        ans=float(dumx)/float(dum0)
        ans=ans*100
    except:
        return 0
    else:
        return ans
def find_category(x):
    try:
        Total_matches=kp0.extract_keywords(x)
        y0=len(Total_matches)
        n=len(x)
        per=percentage1(n,y0)
        if y0==0:
            matched_catkeywords='None'
        matched_catkeywords=Total_matches
        return y0,matched_catkeywords,per
    except:
        print(x)
Total=[]
words=[]
per=[]
df1=data
for i in tqdm(df1['cleaned_website_text']):
    Total.append(find_category(i)[0])
    words.append(find_category(i)[1])
    per.append(find_category(i)[2])
df1['no of matched keywords']=Total
df1['words matched']=words
df1['percent']=per
cat=[]
for i in per:
    if i>0:
        cat.append("drug")
    else:
        cat.append("non drug")
    
df1['category']=cat
df1['category'].value_counts()

# df.to_csv("datawithsmote.csv")

from flashtext.keyword import KeywordProcessor
kp0=KeywordProcessor()
c=0
for word in cossmote:
    kp0.add_keyword(word)
    c=c+1
print(c)
def percentage1(dum0,dumx):
    try:
        ans=float(dumx)/float(dum0)
        ans=ans*100
    except:
        return 0
    else:
        return ans
def find_category(x):
    try:
        Total_matches=kp0.extract_keywords(x)
        y0=len(Total_matches)
        n=len(x)
        per=percentage1(n,y0)
        if y0==0:
            matched_catkeywords='None'
        matched_catkeywords=Total_matches
        return y0,matched_catkeywords,per
    except:
        print(x)
Total=[]
words=[]
per=[]
df2=data
for i in tqdm(df2['cleaned_website_text']):
    Total.append(find_category(i)[0])
    words.append(find_category(i)[1])
    per.append(find_category(i)[2])
df2['no of matched keywords']=Total
df2['words matched']=words
df2['percent']=per
cat=[]
for i in per:
    if i>0:
        cat.append("drug")
    else:
        cat.append("non drug")
    
df2['category']=cat
df2['category'].value_counts()

# df.to_csv("datawithcossmote.csv")

from flashtext.keyword import KeywordProcessor
kp0=KeywordProcessor()
c=0
for word in cleanog:
    kp0.add_keyword(word)
    c=c+1
print(c)
def percentage1(dum0,dumx):
    try:
        ans=float(dumx)/float(dum0)
        ans=ans*100
    except:
        return 0
    else:
        return ans
def find_category(x):
    try:
        Total_matches=kp0.extract_keywords(x)
        y0=len(Total_matches)
        n=len(x)
        per=percentage1(n,y0)
        if y0==0:
            matched_catkeywords='None'
        matched_catkeywords=Total_matches
        return y0,matched_catkeywords,per
    except:
        print(x)
Total=[]
words=[]
per=[]
df3=data
for i in tqdm(df3['cleaned_website_text']):
    Total.append(find_category(i)[0])
    words.append(find_category(i)[1])
    per.append(find_category(i)[2])
df3['no of matched keywords']=Total
df3['words matched']=words
df3['percent']=per
cat=[]
for i in per:
    if i>0:
        cat.append("drug")
    else:
        cat.append("non drug")
    
df3['category']=cat
df3['category'].value_counts()

df3.to_csv("datawithcleanog.csv")
