==> 4. Validation.ipynb <==
import pandas as pd
import numpy as np
from scipy import spatial
import matplotlib.pyplot as plt
from sklearn.manifold import TSNE
import smote_variants as sv
from tqdm import tqdm
from urllib.parse import urlparse
import requests
from collections import Counter
import pandas as pd
import os

df1=pd.read_csv("dataset/datawithoglist.csv")
df2=pd.read_csv("dataset/datawithcleanog.csv")
df3=pd.read_csv("dataset/datawithsmote.csv")
df4=pd.read_csv("dataset/datawithcossmote.csv")
df1=df1[['body','description','title','url','cleaned_website_text','words matched','category']]
df1.rename(columns = {'category':'cat_og'}, inplace = True)
df1.rename(columns = {'words matched':'words_og'}, inplace = True)
df2=df2[['body','description','title','url','cleaned_website_text','words matched','category']]
df2.rename(columns = {'category':'cat_ogclean'}, inplace = True)
df2.rename(columns = {'words matched':'words_ogclean'}, inplace = True)
df3=df3[['body','description','title','url','cleaned_website_text','words matched','category']]
df3.rename(columns = {'category':'cat_smote'}, inplace = True)
df3.rename(columns = {'words matched':'words_smote'}, inplace = True)
df4=df4[['body','description','title','url','cleaned_website_text','words matched','category']]
df4.rename(columns = {'category':'cat_cossmote'}, inplace = True)
df4.rename(columns = {'words matched':'words_cossmote'}, inplace = True)

df1['cat_og'].value_counts()

df2['cat_ogclean'].value_counts()

df3['cat_smote'].value_counts()

df4['cat_cossmote'].value_counts()

cat_ogclean=list(df2['cat_ogclean'])
cat_smote=list(df3['cat_smote'])
cat_cossmote=list(df4['cat_cossmote'])
words_ogclean=list(df2['words_ogclean'])
words_smote=list(df3['words_smote'])
words_cossmote=list(df4['words_cossmote'])
df1['words_ogclean']=words_ogclean
df1['cat_ogclean']=cat_ogclean
df1['words_smote']=words_smote
df1['cat_smote']=cat_smote
df1['words_cossmote']=words_cossmote
df1['cat_cossmote']=cat_cossmote
df12=df1.drop_duplicates(keep="first",subset=['url'])
df12

test1=df1.query('cat_og=="non drug" & cat_ogclean=="non drug"')
len(test1)

test2=df1.query('cat_og=="drug" & cat_ogclean=="non drug"')
len(test2)

test3=df1.query('cat_og=="non drug" & cat_ogclean=="drug"')
len(test3)

#new testcases

t1=df1.query('cat_og=="non drug" & cat_cossmote=="non drug"')
len(t1)

t2=df1.query('cat_og=="drug" & cat_cossmote=="non drug"')
len(t2)

t3=df1.query('cat_og=="non drug" & cat_cossmote=="drug"')
len(t3)

t3.head(5)

import ast
l1=[]
l2=[]
urls=['http://c5xoy22aadb2rqgw3jh2m2irmu563evukqqddu5zjandunaimzaye5id.onion/',
      'http://gkcns4d3453llqjrksxdijfmmdjpqsykt6misgojxlhsnpivtl3uwhqd.onion/',
     'http://porf65zpwy2yo4sjvynrl4eylj27ibrmo5s2bozrhffie63c7cxqawid.onion/',
     'http://4p6i33oqj6wgvzgzczyqlueav3tz456rdu632xzyxbnhq4gpsriirtqd.onion/',
     'http://gn74rz534aeyfxqf33hqg6iuspizulmvpd7zoyz7ybjq4jo3whkykryd.onion/']
x=df1.loc[df1['url']==urls[0]] 
l1.append(set(ast.literal_eval(x['words_og'].iloc[1])))
l2.append(set(ast.literal_eval(x['words_cossmote'].iloc[1])))
x=df1.loc[df1['url']==urls[1]] 
l1.append(set(ast.literal_eval(x['words_og'].iloc[2])))
l2.append(set(ast.literal_eval(x['words_cossmote'].iloc[2])))
x=df1.loc[df1['url']==urls[2]] 
l1.append(set(ast.literal_eval(x['words_og'].iloc[1])))
l2.append(set(ast.literal_eval(x['words_cossmote'].iloc[1])))
x=df1.loc[df1['url']==urls[3]] 
l1.append(set(ast.literal_eval(x['words_og'].iloc[2])))
l2.append(set(ast.literal_eval(x['words_cossmote'].iloc[2])))
x=df1.loc[df1['url']==urls[4]] 
l1.append(set(ast.literal_eval(x['words_og'].iloc[0])))
l2.append(set(ast.literal_eval(x['words_cossmote'].iloc[0])))

l1

l2
