==> 8. News Tagging.ipynb <==
<H2> GENERATE SYNTHETIC WORDS</H2>

import numpy as np
from scipy import spatial
import matplotlib.pyplot as plt
from sklearn.manifold import TSNE
import pandas as pd
import random
from sklearn.neighbors import NearestNeighbors
from sklearn.manifold import TSNE as tsne
from sklearn.decomposition import PCA
import plotly.express as px
from numpy.linalg import norm
from tqdm import tqdm
import spacy as sp
from collections import Counter
import en_core_web_sm
import re
import pickle

with open('glovedata.pickle', 'rb') as handle:
    data_dict = pickle.load(handle)

slist=["development","java","algorithms","computer","laptop","technical","programming","technology","code","debuggung",
       "software","android","device","byte","hardware","digital","browser",'games',
       "cybersecurity","tech","desktop","machine-learning","artificial-inteligence",
       "microprocessor","compiler","database","cloud","graphics","microsoft"]
category="computers"

vectors=[]
x=[]
words=[]
for i in slist:
    try:
        vectors.append(data_dict[i])
        words.append(i)
    except:
        x.append(i)
c=[]
error=[]
for i in x:
    s=i.split(" ")
    for j in s:
        c.append(j)
for i in c:
    try:
        vectors.append(data_dict[i])
        words.append(i)
    except:
        error.append(i)
print("Preprocessed words: ",len(words))
print(words)

def random_sample(vectors):
    x=random.choice(vectors)
    return x
def knn(vec):
    nbs=NearestNeighbors(n_neighbors=5,metric='euclidean',algorithm='kd_tree').fit(vec)
    euclidean,indices= nbs.kneighbors(vec)
    return indices,euclidean
def random_number():
    return np.random.random(1)[0]

def find_closest_embeddings(embedding):
    return sorted(data_dict.keys(), key=lambda word: spatial.distance.euclidean(data_dict[word], embedding))

index,dist=knn(vectors)
random=random_number()
newwords=[]
for i in tqdm(index):
    newsamp=(vectors[i[0]]-vectors[i[1]])*random
    finalsamp=vectors[i[0]]+abs(newsamp)
    l1=find_closest_embeddings(finalsamp)[0:15]
    newwords.append(l1)

flist=[]
for i in newwords:
    for j in i:
        if j not in flist and j not in words:
            flist.append(j)
len(flist)

def cosine_similarity(a, b):
    nominator = np.dot(a, b)
    a_norm = np.sqrt(np.sum(a**2))
    b_norm = np.sqrt(np.sum(b**2))
    denominator = a_norm * b_norm
    cosine_similarity = nominator / denominator
    return cosine_similarity

count=0
a1=[]
a2=[]
for i in flist:
#     print(i,cosine_similarity(data_dict[category[0]],data_dict[i]),cosine_similarity(data_dict[category[1]],data_dict[i]))
    if cosine_similarity(data_dict[category],data_dict[i])>0.5:
        a2.append(i)

a2

def find_closest_embeddings(embedding):
    return sorted(data_dict.keys(), key=lambda word: spatial.distance.euclidean(data_dict[word], embedding))
newlist=[]
for i in a2:
    near=(find_closest_embeddings(data_dict[i])[0:15])
    newlist.append(near)

flist=[]
for i in newlist:
    for j in i:
        if j not in flist:
            flist.append(j)
len(flist)

def find_closest_embeddings(embedding):
    return sorted(data_dict.keys(), key=lambda word: spatial.distance.euclidean(data_dict[word], embedding))
newlist=[]
for i in tqdm(flist):
    near=(find_closest_embeddings(data_dict[i])[0:5])
    newlist.append(near)

for i in newlist:
    for j in i:
        if cosine_similarity(data_dict[category],data_dict[j])>0.15:
#             
            if j not in flist:
                flist.append(j)
len(flist)

fflist=[]
for i in flist:
    if cosine_similarity(data_dict[category],data_dict[i])>0.3:
        fflist.append(i)
len(fflist)

for i in fflist:
    if i not in slist:
        slist.append(i)
len(slist)

# for i in flist:
#     slist.append(i)
# len(slist)

slist

<H2>LABEL DATASET</H2>

print("Number of synthetic words generated using improvised smote ",len(slist))

sp.prefer_gpu()
nlp = en_core_web_sm.load()
def clean_text(doc):
    '''
    Clean the document. Remove pronouns, stopwords, lemmatize the words and lowercase them
    '''
    doc = nlp(doc)
    tokens = []
    exclusion_list = ["nan"]
    for token in doc:
        if token.is_stop or token.is_punct or token.text.isnumeric() or (token.text.isalnum()==False) or token.text in exclusion_list :
            continue
        token = str(token.lemma_.lower().strip())
        tokens.append(token)
    return " ".join(tokens) 

data=pd.read_csv("C:/Users/hetul/Downloads/bbc-text.csv")
data["cleaned_text"]  = data.apply(lambda row: clean_text(row["text"]) ,axis=1)

data.replace(to_replace ="sport", value = "not technical", inplace = True)
data.replace(to_replace ="business", value = "not technical",  inplace = True)
data.replace(to_replace ="politics", value = "not technical", inplace = True)
data.replace(to_replace ="entertainment",  value = "not technical",  inplace = True)
data

data.category.value_counts()

data.isnull().sum()

tech=pd.read_csv("techwords.csv")
wrds=tech['words'].tolist()
print("Total synthetic words generated: ",len(wrds))
print(wrds[:100])

from flashtext.keyword import KeywordProcessor
kp0=KeywordProcessor()
c=0
for word in wrds:
    kp0.add_keyword(word)
    c=c+1
print(c)
def percentage1(dum0,dumx):
    try:
        ans=float(dumx)/float(dum0)
        ans=ans*100
    except:
        return 0
    else:
        return ans
def find_category(x):
    try:
        Total_matches=kp0.extract_keywords(x)
        y0=len(Total_matches)
        n=len(x.split())
        per=percentage1(n,y0)
        if y0==0:
            matched_catkeywords='None'
        matched_catkeywords=Total_matches
        return y0,matched_catkeywords,per
    except:
        print(x)
Total=[]
words=[]
per=[]
df2=data
for i in tqdm(df2['cleaned_text']):
    Total.append(find_category(i)[0])
    words.append(find_category(i)[1])
    per.append(find_category(i)[2])
df2['no of matched keywords']=Total
df2['words matched']=words
df2['percent']=per
cat=[]
for i in per:
    if i>7:
        cat.append("technology")
    else:
        cat.append("not technology")
    
df2['category_predicted']=cat
df2['category_predicted'].value_counts()

x=df2.query('category_predicted=="not technology" & category=="tech"')
x

x=df2.query('category_predicted=="technology" & category=="tech"')
x

df2

df2.to_csv("dataset/scsmotetech.csv")

# pd.DataFrame(slist,columns=["words"]).to_csv("techwords.csv")

data

data.to_csv("dataset/techdata.csv")
